+++ examples/pipelines/unstructured_to_sql_on_the_fly/postgres/init-db.sql
@@ -1,0 +1,10 @@
+CREATE TABLE IF NOT EXISTS quarterly_earnings (
+    company_symbol TEXT NOT NULL,
+    eps FLOAT NOT NULL,
+    net_income_md FLOAT NOT NULL,
+    quarter TEXT NOT NULL,
+    revenue_md FLOAT NOT NULL,
+    year BIGINT NOT NULL,
+    time BIGINT NOT NULL,
+    diff INTEGER NOT NULL
+);